Refactor(LottoStore): 비슷한 역할 클래스끼리 통합
